+++ SapphireXR_App/Models/PLCService/PLCService.Read.cs
@@ -77,9 +77,9 @@ namespace SapphireXR_App.Models
                 int iterlock1 = Ads.ReadAny<int>(hInterlock[0]);
                 dLogicalInterlockStateIssuer?.Publish(new BitArray(BitConverter.IsLittleEndian == true ? BitConverter.GetBytes(iterlock1) : BitConverter.GetBytes(iterlock1).Reverse().ToArray()));
 
-                temperatureTVPublisher?.Publish(Ads.ReadAny<short>(hTemperatureTV));
-                pressureTVPublisher?.Publish(Ads.ReadAny<short>(hPressureTV));
-                rotationTVPublisher?.Publish(Ads.ReadAny<short>(hRotationTV));
+                temperatureTVPublisher?.Publish(Ads.ReadAny<float>(hTemperatureTV));
+                pressureTVPublisher?.Publish(Ads.ReadAny<float>(hPressureTV));
+                rotationTVPublisher?.Publish(Ads.ReadAny<float>(hRotationTV));
 
                 foreach (Action task in AddOnPLCStateUpdateTask)
                 {
